+++ asyncgit/src/sync/branch/mod.rs
@@ -252,7 +252,7 @@ pub fn checkout_branch(
         if let Err(e) = repo.checkout_head(Some(
             git2::build::CheckoutBuilder::new().force(),
         )) {
-            // This is safe beacuse cur_ref was just found
+            // This is safe because cur_ref was just found
             repo.set_head(
                 bytes2string(cur_ref.name_bytes())?.as_str(),
             )?;
@@ -264,7 +264,7 @@ pub fn checkout_branch(
     }
 }
 
-///
+/// Modify HEAD to point to a branch then checkout head, does not work if there are uncommitted changes
 pub fn checkout_remote_branch(
     repo_path: &str,
     branch: &BranchInfo,
@@ -290,18 +290,42 @@ pub fn checkout_remote_branch(
     };
 
     let commit = repo.find_commit(branch.top_commit.into())?;
-    let mut new_branch = repo.branch(&name, &commit, false)?;
-    new_branch.set_upstream(Some(&branch.name))?;
+
+    // find existing local branch for the remote branch
+    let branch_to_checkout = if let Ok(matched_local_branch) = repo.find_branch(&name, BranchType::Local) {
+
+        let local_branch_upstream_name = matched_local_branch.upstream()?.name()?.unwrap().to_string();
+
+        // check if the requested branch's full name matches the local's upstream
+        if branch.name == local_branch_upstream_name {
+            matched_local_branch
+        } else {
+            return Err(
+                Error::Generic(
+                    format!(
+                        "A local branch exists with the name {}, but a different upstream: {} instead of {}",
+                        matched_local_branch.name()?.unwrap(),
+                        local_branch_upstream_name,
+                        branch.name
+                    )
+                )
+            )
+        }
+    } else {
+        let mut new_branch = repo.branch(&name, &commit, false)?;
+        new_branch.set_upstream(Some(&branch.name))?;
+        new_branch
+    };
 
     repo.set_head(
-        bytes2string(new_branch.into_reference().name_bytes())?
+        bytes2string(branch_to_checkout.into_reference().name_bytes())?
             .as_str(),
     )?;
 
     if let Err(e) = repo.checkout_head(Some(
         git2::build::CheckoutBuilder::new().force(),
     )) {
-        // This is safe beacuse cur_ref was just found
+        // This is safe because cur_ref was just found
         repo.set_head(bytes2string(cur_ref.name_bytes())?.as_str())?;
         return Err(Error::Git(e));
     }